[fix] HttpRequestParser 일부 메서드 추가 및 반영
